+++ app/src/main/java/ru/itmo/travelhelper/screens/flightScreens/FlightArrivalFragment.kt
@@ -8,9 +8,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
-import ru.itmo.domain.models.flightTicketListModels.AirportModel
-import ru.itmo.domain.models.flightTicketListModels.CityModel
-import ru.itmo.domain.models.flightTicketListModels.CountryModel
+import ru.itmo.domain.models.flightModels.AirportModel
+import ru.itmo.domain.models.flightModels.CityModel
+import ru.itmo.domain.models.flightModels.CountryModel
 import ru.itmo.travelhelper.databinding.FragmentFlightLocationArrivalBinding
 import ru.itmo.travelhelper.presenter.flightPresentors.FlightPresenterArrival
 import ru.itmo.travelhelper.view.flightViews.FlightArrivalFragmentView
