--- domain/src/main/java/ru/itmo/domain/models/flightTicketListModels/CityModel.kt
@@ -1,5 +1,0 @@
-package ru.itmo.domain.models.flightTicketListModels
-
-data class CityModel(
-    val cityName: String,
-)
