--- domain/src/main/java/ru/itmo/domain/usecases/flightTicketListUseCases/GetCityUseCase.kt
@@ -1,11 +1,0 @@
-package ru.itmo.domain.usecases.flightTicketListUseCases
-
-import ru.itmo.domain.models.flightTicketListModels.CityModel
-import ru.itmo.domain.models.flightTicketListModels.CountryModel
-import ru.itmo.domain.repositories.flightTicketListRepositories.LocationRepository
-
-class GetCityUseCase(private val locationRepository: LocationRepository) {
-    suspend fun execute(): Map<CountryModel, List<CityModel>> {
-        return locationRepository.getCities()
-    }
-}
